pyproject.toml
[project]
name = "rag-book"
version = "0.1.0"
description = "A simple RAG Agentic Workflow for Book data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "ipykernel>=6.29.5",
    "llama-index>=0.12.43",
    "llama-index-readers-file>=0.4.9",
    "llama-parse>=0.6.34",
    "openai>=1.91.0",
    "pip>=25.1.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.2",
    "scrapfly-sdk>=0.8.23",
]
